LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletnode-channel-count.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'gainWorkletNode.parameters.get('gain').value = 1.0')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'gainWorkletNode.parameters.get('gain').value = 1.0')

PASS # AUDIT TASK RUNNER STARTED.
TIMEOUT Executing "setup-buffer-and-worklet" Test timed out
NOTRUN Executing "verify-rendered-buffer"
NOTRUN Audit report
PASS > [setup-buffer-and-worklet]
